refactor(knockdog): searchSnapshot url state과 search url state 통합
